requirements.txt
Django==5.2.4
openai==1.30.1
python-dotenv==1.0.1
requests==2.31.0
PyMuPDF==1.24.1
python-docx==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-decouple==3.8
fastapi[standard]==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
python-magic==0.4.27
dj-database-url==2.1.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
sentry-sdk==1.32.0

